+++ aea/helpers/async_utils.py
@@ -649,7 +649,7 @@ class Runnable(ABC):
     def stop(self, force: bool = False) -> None:
         """Stop runnable."""
         _default_logger.debug(f"{self} is going to be stopped {self._task}")
-        if not self._task or not self._loop:
+        if not self._task or not self._loop:  # pragma: nocover
             return
 
         if self._task.done():
